feat: ssh: Add output_mode and output_manual_delimiters parameters.
output_mode allow template editor to choose how output will be retrieve
when using a SSH step.

output_delimiters is an array parameter to personalize where uTask
should look for an output. If the command doesn't return a JSON output
on the last line, it's possible to give a start/end pattern for uTask to
retrieve the output serialized on JSON-format. output_delimiters should
be an array with two elements, start and end pattern.

Signed-off-by: Romain Beuque <romain.beuque@corp.ovh.com>