feat: add tags to a task and an associated builtin plugin
This adds a a new metadata field to the Task type: tags. Tags can be used
to filter out specific tasks. The filtering is restricted to exact key/value
match for now.

A builtin plugin called "tag" can be used to update the tags of a task during
its execution. Note that no guarantees are made if two steps of the same task
running in parallel update the task. For the time being, the tag builtin plugin
should be used sequentially with others steps.

Signed-off-by: William Poussier <william.poussier@gmail.com>
Co-Authored-By: Thomas Schaffer <thomas.schaffer@corp.ovh.com>