fix: runAvailableSteps should only look at expandedSteps on the current loop
We are injecting expandedSteps during a loop to the next call of
runAvailableSteps to add them as candidates in the next call.

If the expandedStep has a SKIP condition, putting the state to a
Runnable one ; then not resetting the expandedSteps array would made
uTask fall into an infinite loop.

Signed-off-by: Romain Beuque <romain.beuque@corp.ovh.com>