fix: foreach: Prevent unwanted children to be commited back to resolution after contract
When all children of a foreach steps are in PRUNE steps, the parent
foreach step will become available for contraction before the step
pop in the stepChan inside resolve().

If parent step is contracted and commited before all occurences of
children being resolve, those occurences will be commited after contraction
and will be persisted in the step tree, which we don't want.

Signed-off-by: Romain Beuque <romain.beuque@corp.ovh.com>