fix(engine): foreach steps in an unexpected state would block the resolution indefinitely
In the case where a foreach step falls into a FATAL_ERROR, or
SERVER_ERROR state, the step was never returned in the stepChan, causing
the resolution to wait indefinitely for this step to finish running.

Adding only failure state of a foreach step into the executedSteps,
otherwise fast-running children of a foreach step would have to be
picked-up for contraction at the next run of the resolution (which would
cause delay for users, and all breaks a lot of tests...).

Signed-off-by: Romain Beuque <romain.beuque@ovhcloud.com>