feat: advanced watcher usernames management
This CL introduces a new plugin `watcher` that allows to add new
watcher usernames to a task, similar to how the `tag` plugin works.

A new field `allowed_watcher_usernames` has also been added to the
core model of a task template to allows pre-defined watchers usernames
at the template level. The pre-defined list of usernames of the template
is merged with the input list given during the creation of a task, ignoring
any duplicate values.

Signed-off-by: William Poussier <william.poussier@gmail.com>