resource limiting: complete rework to handle context cancelation and native rate-limiting on a template
This introduces a complete rework of resource usage by implementing
contexts, so the lock can be cancelled, in case the lock takes too long
to be obtained, or if the instance is shutting-down.

It also introduces a native resource on all resolution based on the
template name: it allows to reduce the concurrent execution of a given
template, or to completely stop the execution of a template if it behave
incorrectly.

Signed-off-by: Romain Beuque <556072+rbeuque74@users.noreply.github.com>