db: check that latest SQL migration is applied before starting µTask
Adding an SQL table that will contains the current µTask SQL migration
that has been applied. This will allow us to make sure that SQL
migration are applied before µTask is started, to make sure we don't
have incompabilities.

Removed support for Go 1.13 & 1.14 as 1.15 is stable since a few months.

Signed-off-by: Romain Beuque <556072+rbeuque74@users.noreply.github.com>