templating: adding fieldFrom helper function to retrieve a field from a previous expression
`field` templating helper function allow to retrieve a key into a
map[string], even if the key is not a valid templating keyword.
`fieldFrom` provides the same feature, expect that it doesn't use the
values map from the task/resolution/steps, but use the previous
templating pipeline as a templating source.

Closes #200

Signed-off-by: Romain Beuque <556072+rbeuque74@users.noreply.github.com>