resolution: allow resolution edition on a single step + add endpoint to change only step state
Adding 3 new API endpoint to retrieve the value of a given step
resolution, edit the content of a given step resolution, and edit the
state of a given step resolution.
This allow a more granular edition, which prevent massive resolution
edition by multiple human operators that would override themselves.

Edition of the state of a given step resolution is allowed to Resolution
Managers, as it would not put any risk to the resolution consistency.

Signed-off-by: Romain Beuque <556072+rbeuque74@users.noreply.github.com>