feat(notification): add new notification type: task_validation
New notification type has been introduced: task_validation. It will be
fired every time a task is created and the task will require to be
validated by a human operator, aka, task will not be auto-runned.
This will concerns tasks that require resolver_inputs, task not declared
as auto_runnable, or task created by a requester which is not an
allowed_resolver.

Signed-off-by: Romain Beuque <556072+rbeuque74@users.noreply.github.com>