collector: switch to SKIP LOCKED (#270)
When µTask is used with a lot of tasks, lock on the postgres can take
a lot of memory as the `pg_try_advisory_xact_lock` lock function will
try to lock all the rows. As Postgres 9.5+ now handle `SKIP LOCKED`
keyword, it can be used to try to acquire only one row, without locking
all the existing resolution. The lock will also be scopped at the table
level, instead of the Postgres cluster level previously.

Signed-off-by: Romain Beuque <556072+rbeuque74@users.noreply.github.com>