fix: race condition inside an expanded step regarding dependencies
Children steps Dependencies array was shared between all the child steps
and the parent step, every modification to one child steps was
propagated to the other child. Impact was detected on template steps
with `foreach_strategy` set to `sequence` as the child steps might be
launched all together, instead of waiting the previous one.

Signed-off-by: Romain Beuque <556072+rbeuque74@users.noreply.github.com>