fix: failure to commit leads to inconsistencies between workflow and database
When a µTask worker fails to commit its current state to the database,
it was sometimes not stopping the execution of its resolution. This
leads to inconsistencies, as the step was currently RUNNING, but
database state indicates TODO. In that way, if the worker crashes, then
the new worker will restart with wrong information, inducing workflow
issues.

Signed-off-by: Romain Beuque <556072+rbeuque74@users.noreply.github.com>