fix: rbac: update of a step state is reserved to template owners only (#307)
Previously, edition of the step state was allowed to resolvers. But
resolvers are regular users, that have the right to resolve the task,
they are not aware of the behaviour of the template. This power should
be reserved to template owners, who wrote the template, and know how the
steps state can be changed.

Signed-off-by: Romain Beuque <556072+rbeuque74@users.noreply.github.com>