feat: resolution: implemented task start-over (#308)
Once a resolution is executed, we can now start-over a task to its
initial state if the resolution were stuck in BAD_REQUEST, CANCELLED or
PAUSED state. It grants admins a way to amend the execution of a
resolution. This feature can also be granted to resolvers of a
resolution, so they can start-over themselves, using the boolean
`allow_task_start_over` in the template.

Signed-off-by: Romain Beuque <556072+rbeuque74@users.noreply.github.com>