feat(conditions): allow running check on retryable states
This allows running check conditions on retryable states
such as SERVER_ERROR, effectively giving better control to the
execution flow.

A use case may be to retry only a couple times on specific non-critical
steps.

This is a non-breaking change, the feature is only enable if a condition
bears the specific boolean.

Signed-off-by: Gabriel Augendre <gabriel.augendre@ovhcloud.com>