Add email attachment support to the Email action
Adds the ability to attach files to emails. This is useful if a previous
step generates a report that you then want to email out.

I extracted out the building of `*mail.Message` to make the code more
testable without needing a fake SMTP server. I opted to test for various
headers that we would expect given our input.

Documentation updated to cover the new parameter.

Signed-off-by: Brian Turchyn <brian@dj-bri-t.net>